Change type of the length parameter in merge sort from int to size_t.

+++ c/algorithms/merge_sort_bottom_up.h
@@ -1,7 +1,8 @@
 #ifndef MERGE_SORT_BOTTOM_UP_H
 #define MERGE_SORT_BOTTOM_UP_H
 
-// TODO(michaelreneer): Change type of length parameter from int to size_t.
-void merge_sort_bottom_up(int array[], int length);
+#include <stddef.h>
+
+void merge_sort_bottom_up(int array[], size_t length);
 
 #endif  // MERGE_SORT_BOTTOM_UP_H
